Migrated to XCTest

+++ NSString+Ruby/NSString+Ruby.h
@@ -19,22 +19,22 @@
 
 //Operator-likes
 -(NSString*):(NSObject*)concat, ...;
--(NSString*)x:(int)mult;
+-(NSString*)x:(NSInteger )mult;
 
 //Shorthand Accessors
--(NSString*):(int)loc :(int)len;
--(NSString*):(int)start :(char*)shorthand :(int)end;
+-(NSString*):(NSInteger )loc :(NSInteger )len;
+-(NSString*):(NSInteger )start :(char*)shorthand :(NSInteger )end;
 
 //Ruby Methods
 -(void)bytes:(void(^)(unichar))block;
-- (NSString*)center:(int)amount;
-- (NSString*)center:(int)amount with:(NSString*)padString;
+- (NSString*)center:(NSInteger )amount;
+- (NSString*)center:(NSInteger )amount with:(NSString*)padString;
 -(void)chars:(void(^)(unichar))block;
 -(NSString*)chomp;
 -(NSString*)chomp:(NSString*)string;
 -(NSString*)chop;
 -(NSString*)chr;
--(void)codePoints:(void(^)(int))block;
+-(void)codePoints:(void(^)(NSInteger ))block;
 - (NSString*)concat:(id)concat;
 -(NSInteger)count:(NSString*)setString, ...;
 -(NSString*)delete:(NSString*)first, ...;
@@ -42,31 +42,31 @@
 -(long)hex;
 -(BOOL)includes:(NSString*)include;
 -(NSInteger)index:(NSString*)pattern;
--(NSInteger)index:(NSString*)pattern offset:(int)offset;
--(NSString*)insert:(int)index string:(NSString*)string;
+-(NSInteger)index:(NSString*)pattern offset:(NSInteger )offset;
+-(NSString*)insert:(NSInteger)index string:(NSString*)string;
 -(NSString*)inspect;
 -(BOOL)isASCII;
 -(BOOL)isEmpty;
 -(NSInteger)lastIndex:(NSString*)pattern;
--(NSInteger)lastIndex:(NSString*)pattern offset:(int)offset;
--(NSString*)leftJustify:(int)amount;
--(NSString*)leftJustify:(int)amount with:(NSString*)padString;
+-(NSInteger)lastIndex:(NSString*)pattern offset:(NSInteger )offset;
+-(NSString*)leftJustify:(NSInteger )amount;
+-(NSString*)leftJustify:(NSInteger )amount with:(NSString*)padString;
 -(NSString*)leftStrip;
 -(void)lines:(void(^)(NSString*))block;
 -(void)lines:(void(^)(NSString*))block separator:(NSString*)separator;
 -(NSArray*)match:(NSString*)pattern;
--(NSArray*)match:(NSString*)pattern offset:(int)offset;
+-(NSArray*)match:(NSString*)pattern offset:(NSInteger )offset;
 -(NSInteger)occurencesOf:(NSString*)subString;
 -(long)octal;
--(int)ordinal;
+-(NSInteger )ordinal;
 -(NSArray*)partition:(NSString*)pattern;
 -(NSString*)prepend:(NSString*)prefix;
 -(NSRange)range;
 -(NSString*)reverse;
 -(NSInteger)rightIndex:(NSString*)pattern;
--(NSInteger)rightIndex:(NSString*)pattern offset:(int)offset;
--(NSString*)rightJustify:(int)amount;
--(NSString*)rightJustify:(int)amount with:(NSString*)padString;
+-(NSInteger)rightIndex:(NSString*)pattern offset:(NSInteger )offset;
+-(NSString*)rightJustify:(NSInteger )amount;
+-(NSString*)rightJustify:(NSInteger )amount with:(NSString*)padString;
 -(NSArray*)rightPartition:(NSString*)pattern;
 -(NSString*)rightStrip;
 - (NSArray*)scan:(NSString*)pattern;
@@ -74,15 +74,15 @@
 - (NSString*)strip;
 - (NSArray*)split;
 - (NSArray*)split:(NSString*)pattern;
-- (NSArray*)split:(NSString*)pattern limit:(int)limit;
+- (NSArray*)split:(NSString*)pattern limit:(NSInteger )limit;
 - (NSString*)squeeze;
 - (NSString*)squeeze:(NSString*)pattern;
 - (NSString*)substituteFirst:(NSString*)pattern with:(NSString*)sub;
 - (NSString*)substituteLast:(NSString*)pattern with:(NSString*)sub;
 -(NSString*)substituteAll:(NSDictionary*)subDictionary;
 -(NSString*)substituteAll:(NSString*)pattern with:(NSString*)sub;
--(int)sum;
--(int)sum:(int)bit;
+-(NSInteger )sum;
+-(NSInteger )sum:(NSInteger )bit;
 -(NSString*)swapCase;
 
 //Subscript Protocol
